add likes table

+++ database/migrations/2020_12_01_200940_create_posts_table.php
@@ -14,8 +14,7 @@ class CreatePostsTable extends Migration
     public function up()
     {
         Schema::create('posts', function (Blueprint $table) {
-            $table->increments('id');
-            // $table->unsignedBigInteger('user_id');
+            $table->bigIncrements('id');
             $table->unsignedBigInteger('user_id')->nullable();
             $table->string('title', 50);
             $table->text('body');
